Adding Server Invite to h>support and Fixing Bugs
In both h>say and h>upload, messages and level names, respectively, had no spaces between them. This is now fixed.
h>support now gives an invitation to the Discord server instead of returning an error.
Toxi accidentally put a . instead of a : in some places. This broke the code that kept the bot running (not the actual code). This is now fixed.